@@ -9,6 +9,6 @@ namespace GhostUI.Models
         public string Summary       { get; set; }
 
         public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-        public int ID => Convert.ToInt32(DateFormatted.Replace("/", ""));
+        public int Id => Convert.ToInt32(DateFormatted.Replace("/", ""));
     }
 }
